messaging_app/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
Markdown==3.8
sqlparse==0.5.3
